requirements.txt
# editor
pylint==2.4.4
autopep8==1.5
pylint_flask_sqlalchemy==0.2.0

# project
flask==1.1.
flask-sqlalchemy==2.4.1
flask-migrate==2.5.2
flask-cors==3.0.8
gunicorn==20.0.4
python-dotenv==0.11.0
